Move getValueForMarker method to Parser class
